Web Track
Basic unerstanding of Data communication
basic understanding of python
basic understanding of flask

Introduction

How to send message over the internate?
what are the languages involved when designing web applications on the internate, HTML, CSS, Javascript, Python, flask...

HTML a language use to design web pages that the web browser can understand.
CSS styles that wabpage.
Javascript more like HTML and CSS we can add more interaction, respond to users input and other ways of making the web pages more dynamic for the users.
Flask framework a pythone framework that allow us to create a web applications that are interactive that are able to store informations and more by combining tools from Python and SQL.

Python + SQL =>we can create user interactive like log in log out list items and more...

How do we connect computers to one another, how do we transfer messages, datas over the internate

To strat with we use TCP/IP protocols

TCP Transmtion control protocol
IP Internate protocol
Thus control how commucation occure throught the internate
Just like sending an postmail with senders and reciever we have IP address for addressing each computer or node's
IP address  #.#.#.# this # each have number that ranges 0-255,0-255, 0-255,0-255-> that is each of this numbers are reprsented by 8 bits of information
hence they are in total 32 bits meanig there can only be 4 billion addresses that are availble on the internte that are connected to eachother with address to which information or message can be sent or recieved.

4 billions is not enough to cover all the nodes available on the internate this is names as IPV4. therfore there exists a newer version IPV6 that uses a 128-bit addresses allowing for far more addresses
that can be availble on the internate.

A computer must be aware of what kind of information they are reciving it may email, web pages, file transfer packets system this is other information computer should be aware of it is vital that the computer knows the relevancy of the file
SO each of these services like email, web pages or file transfer in order to solve this problem we assign a port number
some common port number are 21 for FTP: file transform protocol, 25 for SMTP commonl used for email, 80 for HTTP for web pages......
Example 1.2.3.4:80 IP address 1.2.3.4, port number 80 a wep page. but this are not using this instead we use URLS http://www.example.com to specfy which web site we want use HOW?

We use DNS
 it is a mapping between URL and IP address
		URL	          IP address
		google.com	  172.217.7.206

Send a request

GET / HTTP /1.1
HOST: www.example.com -> specfying a web page request
...
This slash / says give me the root -> here GET / HTTP / 1.1

Response can be
HTTP/1.1 200 OK
Content-Type: text/html  --> the response content type is text/html
...

Other standard code includes
	200	OK
	301	Moved permanently; you may be redirected to other route
	403	Forbiden    ; You don't have permition to see this
	404	Not found   ; it is just not found
	500	Internal server error

Symplified search method web pages can take web parameters a placeholder to specfiy here is additional information about the request

JavaScript

We can see the DOM structure of html how it is constructed at each step yet by using javascript we can manipulate the DOM and do things with it
to create interactive web pages and add more functionalties.

Web Track

Basic unerstanding of Data communication
basic understanding of python
basic understanding of flask

Introduction

How to send message over the internate?
what are the languages involved when designing web applications on the internate, HTML, CSS, Javascript, Python, flask...

HTML a language use to design web pages that the web browser can understand.
CSS styles that wabpage.
Javascript more like HTML and CSS we can add more interaction, respond to users input and other ways of making the web pages more dynamic for the users.
Flask framework a pythone framework that allow us to create a web applications that are interactive that are able to store informations and more by combining tools from Python and SQL.

Python + SQL =>we can create user interactive like log in log out list items and more...

How do we connect computers to one another, how do we transfer messages, datas over the internate

To strat with we use TCP/IP protocols

TCP Transmtion control protocol
IP Internate protocol
Thus control how commucation occure throught the internate
Just like sending an postmail with senders and reciever we have IP address for addressing each computer or node's
IP address  #.#.#.# this # each have number that ranges 0-255,0-255, 0-255,0-255-> that is each of this numbers are reprsented by 8 bits of information
hence they are in total 32 bits meanig there can only be 4 billion addresses that are availble on the internte that are connected to eachother with address to which information or message can be sent or recieved.

4 billions is not enough to cover all the nodes available on the internate this is names as IPV4. therfore there exists a newer version IPV6 that uses a 128-bit addresses allowing for far more addresses
that can be availble on the internate.

A computer must be aware of what kind of information they are reciving it may email, web pages, file transfer packets system this is other information computer should be aware of it is vital that the computer knows the relevancy of the file
SO each of these services like email, web pages or file transfer in order to solve this problem we assign a port number
some common port number are 21 for FTP: file transform protocol, 25 for SMTP commonl used for email, 80 for HTTP for web pages......
Example 1.2.3.4:80 IP address 1.2.3.4, port number 80 a wep page. but this are not using this instead we use URLS http://www.example.com to specfy which web site we want use HOW?

We use DNS
 it is a mapping between URL and IP address
		URL	          IP address
		google.com	  172.217.7.206

Send a request

GET / HTTP /1.1
HOST: www.example.com -> specfying a web page request
...
This slash / says give me the root -> here GET / HTTP / 1.1

Response can be
HTTP/1.1 200 OK
Content-Type: text/html  --> the response content type is text/html
...

Other standard code includes
	200	OK
	301	Moved permanently; you may be redirected to other route
	403	Forbiden    ; You don't have permition to see this
	404	Not found   ; it is just not found
	500	Internal server error

Symplified search method web pages can take web parameters a placeholder to specfiy here is additional information about the request

JavaScript

We can see the DOM structure of html how it is constructed at each step yet by using javascript we can manipulate the DOM and do things with it
to create interactive web pages and add more functionalties.